test(server): consolidate reusable test infrastructure
