+++ packages/server/src/testing/index.ts
@@ -1,0 +1,5 @@
+export { Deferred, TEST_MODEL, TEST_TOKEN, TestSessionBackend, TestSessionRuntime } from "./backend.ts";
+export type { WireChannel } from "./client.ts";
+export { connectUnixTestClient, ProtocolTestClient } from "./client.ts";
+export type { TestServer, TestServerOptions } from "./server.ts";
+export { createTestServer } from "./server.ts";
